Restructure build dependencies.
