+++ lib/deploy/lambda/index.js
@@ -1,12 +1,13 @@
+const archiver = require('archiver')
 const fs = require('fs')
 const { join, basename } = require('path')
-const getPreviousResult = require('../../build-if-necessary/get-previous-result')
-const archiver = require('archiver')
+const getPreviousResult = require('../../build/build-if-necessary/get-previous-result')
+const buildResultFilename = require('../../build/build-result-filename')
 
 class LambdaDeploy {
   constructor (dir) {
     this.dir = dir
-    this.brbBuildFile = join(dir, '.brb_build.json')
+    this.brbBuildFile = join(dir, buildResultFilename)
   }
 
   get buildDone () {
@@ -49,7 +50,7 @@ class LambdaDeploy {
     archive.glob('**/*', {
       cwd: join(process.cwd(), this.dir),
       dot: true,
-      ignore: ['.brb_build.json', this.appBasename]
+      ignore: [buildResultFilename, this.appBasename]
     }, { prefix: 'public' })
 
     archive.glob('node_modules/**/*', {
