+++ lib/serve.js
@@ -1,5 +1,5 @@
 const build = require('./build')
-const buildIfNecessary = require('./build-if-necessary')
+const buildIfNecessary = require('./build/build-if-necessary')
 
 const PROD = process.env.NODE_ENV === 'production'
